requirements.txt
streamlit
nltk
wordcloud
pandas
matplotlib
google-cloud-bigquery
google-auth-oauthlib
requests
bs4
regex